Fix create reference evaluator, bounds checks
Change value type argument handling in create reference evaluator to either match the source evaluator, or cast scalars.
https://tracker.physiomeproject.org/show_bug.cgi?id=3835
Fix bounds checks on argument indexes from <0 to <1.
https://tracker.physiomeproject.org/show_bug.cgi?id=3836
Some minor formatting fixes and compilation warnings.

+++ core/src/SimpleMap.h
@@ -97,7 +97,7 @@ public:
 
     int size()
     {
-        return pairs.size();
+        return static_cast<int>(pairs.size());
     }
